requirements.txt
certifi==2020.12.5
chardet==4.0.0
comtypes==1.1.10
idna==2.10
MouseInfo==0.1.3
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.27
pytesseract==0.3.7
pyttsx3==2.90
PyTweening==1.0.3
pywin32==300
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.4
